refactor: extract a command class
