+++ tarentula/download.py
@@ -7,11 +7,12 @@ from requests.exceptions import HTTPError, ConnectionError
 from rich.progress import Progress
 from urllib3.exceptions import ProtocolError
 
+from tarentula.command import Command
 from tarentula.datashare_client import DatashareClient
 from tarentula.logger import logger
 
 
-class Download:
+class Download(Command):
     def __init__(self,
                  datashare_url: str = 'http://localhost:8080',
                  datashare_project: str = 'local-datashare',
@@ -34,9 +35,9 @@ class Download:
                  progressbar: bool = True,
                  raw_file: bool = True,
                  type: str = 'Document'):
+        super().__init__(query, type)
         self.datashare_url = datashare_url
         self.datashare_project = datashare_project
-        self.query = query
         self.destination_directory = destination_directory
         self.throttle = throttle
         self.cookies_string = cookies
@@ -53,7 +54,6 @@ class Download:
         self.size = size
         self.sort_by = sort_by
         self.order_by = order_by
-        self.type = type
         try:
             self.datashare_client = DatashareClient(datashare_url,
                                                     elasticsearch_url,
@@ -63,40 +63,6 @@ class Download:
         except (ConnectionRefusedError, ConnectionError):
             logger.critical('Unable to connect to Datashare', exc_info=self.traceback)
             exit()
-
-    @property
-    def query_body(self):
-        if self.query.startswith('@'):
-            return self.query_body_from_file
-        else:
-            return self.query_body_from_string
-
-    @property
-    def query_body_from_string(self):
-        return {
-            "query": {
-                "bool": {
-                    "must": [
-                        {
-                            "match": {
-                                "type": self.type
-                            }
-                        },
-                        {
-                            "query_string": {
-                                "query": self.query
-                            }
-                        }
-                    ]
-                }
-            }
-        }
-
-    @property
-    def query_body_from_file(self):
-        with open(self.query[1:]) as json_file:
-            query_body = json.load(json_file)
-        return query_body
 
     @property
     def no_progressbar(self):
